Dirty Fix
Fixed various things and commented out colorFade() as it needs more work.

+++ RGBW_controller_and_thermometer/universalFunc.h
@@ -15,7 +15,7 @@ bool isNegative(int _ref) { //If _ref is negative return true
 
 //Timing
 #define countDataAmount 4 //Number of timers
-int countData[countDataAmount] = {0}; //Holds count information. (Adjust for numeber of timers needed.)
+unsigned long countData[countDataAmount] = {0}; //Holds count information. (Adjust for numeber of timers needed.)
 
 
 
@@ -47,4 +47,5 @@ bool timer(int _interval,byte _id) { //_interval in millis, _id in countData
     return false;
   }
 }
+
